Solved some bugs and docker container for redis

+++ Lumina-Ai-Backend/src/main/java/com/lumina/ai/Lumina_Ai_Backend/service/AnonymousService.java
@@ -28,7 +28,7 @@ public class AnonymousService {
             throw new IllegalArgumentException("Prompt Cannot be empty");
         }
         PromptRequest request=new PromptRequest();
-        request.setId(Math.random());
+        request.setId((long)Math.random());
         request.setInput(prompt);
         PromptResponse response=restTemplate.postForObject(aiServiceUrl,request,PromptResponse.class);
         return CompletableFuture.completedFuture(response);
